+++ cmd/root.go
@@ -4,7 +4,7 @@ import (
 	"github.com/spf13/cobra"
 )
 
-const Version = "1.6.0"
+const Version = "1.7.0"
 
 var rootCmd = &cobra.Command{
 	Use:   "snare",